Add files via upload
Within the HTML document you will notice I placed attention to the DOM structure of the html files by using different h1's, h2's, and h3's as well as section header and footer tags for optimal screen reader navigation as well as proficient DOM structure for general SEO including some meta tags for good descriptions. a common mistake is that the variation in h1-h10 is to set your font weight and this is not the case. there is a direct usage for separating the data your browser is rendering although html is a very simple language these are some small details to consider when laying out a good html page.